refactor WebinarCall (add subcomponents)
